Use the function definations in capi

+++ libraries/chain/vm_api/cf_transaction.cpp
@@ -2,12 +2,12 @@
 #include <eosio/vm/span.hpp>
 #include "vm_api_proxy.hpp"
 
-size_t vm_api_proxy::read_transaction(char *data, size_t data_size) {
+uint32_t vm_api_proxy::read_transaction(char *data, size_t data_size) {
     legacy_span<char> _data(data, data_size);
     return _interface->read_transaction(std::move(_data));
 }
 
-size_t vm_api_proxy::transaction_size() {
+uint32_t vm_api_proxy::transaction_size() {
     return _interface->transaction_size();
 }
 
@@ -23,7 +23,7 @@ int32_t vm_api_proxy::tapos_block_prefix() {
     return _interface->tapos_block_prefix();
 }
 
-int32_t vm_api_proxy::get_action( uint32_t type, uint32_t index, char *buffer, size_t buffer_size) {
+int32_t vm_api_proxy::get_action( uint32_t type, uint32_t index, char *buffer, uint32_t buffer_size) {
     legacy_span<char> _buffer(buffer, buffer_size);
     return _interface->get_action(type, index, std::move(_buffer));
 }
